Introduction

The main functionality provided by PlutoDevMacros is the possibility of loading local packages within a Pluto notebook without requiring to activate a local environment (and thus deactivate the Pluto package manager).

This functionality is implemented inside the FromPackage submodule of PlutoDevMacros and accessed throught the @fromparent or @frompackage macros.

Note

Both @fromparent and @frompackage are exported by PlutoDevMacros itself without requiring to explicitly use FromPackage.

While the @fromparent macro was initially developed in order to facilitate creating Julia packages using Pluto notebooks as building blocks, in its current implementation it helps a lot with prototyping and testing during local package development, even when creating normal packages not relying on Pluto notebooks as building blocks.

This macro allows in fact to load the module of a local package within a running Pluto notebook and permits to easily reload the local code upon request similar to a Revise-based workflow but with a few notable advantages:

Package code can be re-evaluated correctly without requiring a julia restart even when re-defining structs or constants

Local code reload, triggered manually via a floating button in the Pluto notebook, automatically triggers execution of all dependent cells, simplifying the process of testing changes of code on specific runtime paths

Possibilty of adding packages to the notebook environment which are not dependencies of the local package, very useful for testing plotting or benchmarking of the local package code without having to put the related packages in either the global or package-local environment

Support for the package extensions functionality added in julia 1.9, which together with the point on notebook environment above simplify the testing and development of extensions on the local package under development.

More details on the synthax and functionality of these macros is given in the following sections.

PlutoDevMacros.FromPackage.@frompackage — Macro

@frompackage target import_block

This macro takes a local Package (derived from the target path, which can be an AbstractString or a @raw_str), loads it as a submodule of the current Pluto workspace and then process the various import/using statements inside import_block to extract varables/functions from the local Package into the notebook workspace.

Its main use is allowing to load a local package under development within a running Pluto notebook in order to facilitate prototyping and testing.

The following julia code inside a Pluto notebook cell:

@frompackage local_package_path begin
	import ^: *
	using >.LocalDependency
end

takes the main module definition code for the package located at local_package_path, creates the corresponding module in the notebook workspace and imports all of the names defined within (That is what the import ^:* statement does).

Additionally, it loads the package called LocalDependency (must be a dependency of the local package) as if the using LocalDependency code was used within the notebook, but without adding LocalDependency to the notebook environment.

See the package documentation for more details.

See also: @fromparent

PlutoDevMacros.FromPackage.@fromparent — Macro

This macro is equivalent to @frompackage but assumes the calling file as the target argument. So the code

@fromparent import_block

is equivalent to

@frompackage @__FILE__ import_block

Refer to the @frompackage docstring and the package documentation for understanding its use. See also: @addmethod