Other Functions

This package also exports some additional convenience macros for simplifying package development aided by Pluto notebooks.

Utilities Macros

PlutoDevMacros.@addmethod — Macro

@addmethod func(args...;kwargs...) = body
@addmethod function func(args...;kwargs...) 
	body
end

This simple macro modifies a function definition expression (only when called from a Pluto notebook) to prepend the name of the module defining the function (here called DefiningModule) to the method definition.

So the code

@addmethod func(args...;kwargs...) = something

is simply translated to

DefiningModule.func(args...;kwargs...) = something

when called from a Pluto notebook, and to:

func(args...;kwargs...) = something

when called outside of Pluto.

This is useful to avoid multiple definition errors inside Pluto but has the caveat that defining a method with @addmethod does not trigger a reactive run of all cells that call the modified function. This also mean that when removing the cell with the @addmethod call, the actual method added to the DefiningModule will not be automatically erased by Pluto and will still accessible until it is not overwritten with another method with the same signature.

This is easy to fix in the case of adding methods to modules loaded with @frompackage/@fromparent as reloading the module is sufficient to remove the hanging method.

See this video for an example:

Link to Video

See also: @frompackage, @fromparent

PlutoDevMacros.@only_in_nb — Macro

only_in_nb(ex)

Executes the expression ex only if the macro is called from a running Pluto instance and ran directly from the source notebook file.

This is more strict than PlutoHooks.@skip_as_script as including a notebook with @skip_as_script ex from another notebook would still execute ex.
@only_in_nb ex instead only evaluates ex if the calling notebook is the original source notebook file.

See also: @only_out_nb. PlutoHooks.@skip_as_script.

PlutoDevMacros.@only_out_nb — Macro

only_out_nb(ex)

Opposite of @only_in_nb

See also: @only_in_nb. PlutoHooks.@only_as_script.

PlutoDevMacros.@current_pluto_cell_id — Macro

@current_pluto_cell_id()

Returns the cell_id (as a string) of the cell where the macro is called. If not ran from within the pluto notebook containing the call, returns an empty string

PlutoDevMacros.@current_pluto_notebook_file — Macro

@current_pluto_notebook_file()

Returns the path of the notebook file of the cell where the macro is called. If not ran from within the pluto notebook containing the call, returns an empty string