Other Functions

This package also exports some additional convenience macros for simplifying package development aided by Pluto notebooks.

Utilities Macros

PlutoDevMacros.@addmethod — Macro

@addmethod func(args...;kwargs...) = body
@addmethod function func(args...;kwargs...) 
	body
end

This simple macro modifies a function definition expression (only when called from a Pluto notebook) to prepend the name of the module defining the function (here called DefiningModule) to the method definition.

So the code

@addmethod func(args...;kwargs...) = something

is simply translated to

DefiningModule.func(args...;kwargs...) = something

when called from a Pluto notebook, and to:

func(args...;kwargs...) = something

when called outside of Pluto.

This is useful to avoid multiple definition errors inside Pluto but has the caveat that defining a method with @addmethod does not trigger a reactive run of all cells that call the modified function. This also mean that when removing the cell with the @addmethod call, the actual method added to the DefiningModule will not be automatically erased by Pluto and will still accessible until it is not overwritten with another method with the same signature.

This is easy to fix in the case of adding methods to modules loaded with @frompackage/@fromparent as reloading the module is sufficient to remove the hanging method.

See this video for an example:

Link to Video

See also: @frompackage, @fromparent

PlutoDevMacros.@only_in_nb — Macro

only_in_nb(ex)

Executes the expression ex only if the macro is called from a running Pluto instance and ran directly from the source notebook file.

This is more strict than PlutoHooks.@skip_as_script as including a notebook with @skip_as_script ex from another notebook would still execute ex.
@only_in_nb ex instead only evaluates ex if the calling notebook is the original source notebook file.

See also: @only_out_nb. PlutoHooks.@skip_as_script.

PlutoDevMacros.@only_out_nb — Macro

only_out_nb(ex)

Opposite of @only_in_nb

See also: @only_in_nb. PlutoHooks.@only_as_script.

PlutoDevMacros.@current_pluto_cell_id — Macro

@current_pluto_cell_id()

Returns the cell_id (as a string) of the cell where the macro is called. If not ran from within the pluto notebook containing the call, returns an empty string

PlutoDevMacros.@current_pluto_notebook_file — Macro

@current_pluto_notebook_file()

Returns the path of the notebook file of the cell where the macro is called. If not ran from within the pluto notebook containing the call, returns an empty string

HTL Script

The Script submodule implements and exports some functionality (built on top of HypertextLiteral.jl) to simplify building javascript content by combining multiple snippets together.

None of the functions/types below are exported by PlutoDevMacros itself, but only by its Script submodule

PlutoDevMacros.Script.HTLScript — Type

Base.@kwdef struct HTLScript
	body::HTLScriptPart
	invalidation::Union{HTLScriptPart, Missing}
	id::Union{Missing, Nothing, String}
end

Fields

body::HTLScriptPart -> The main body of the script

invalidation::Union{HTLScriptPart, Missing} -> The code to be executed inside the invalidation promise. Defaults to missing

id::Union{Missing, Nothing, String} -> The id to assign to the script. Defaults to missing

Additional Constructors

HTLScript(body; kwargs...)
HTLScript(body, invalidation; kwargs...)

For the body and invalidation fields, the constructor also accepts inputs of type String, HypertextLiteral.Result and IOBuffer, translating them into HTLScriptPart internally.

HTLScript(s::HTLScript; kwargs...)

This constructor is used to copy the elements from another HTLScript with the option of overwriting the fields provided as kwargs

Description

This struct is used to create and compose scripts together with the @htl macro from HypertextLiteral.

It is intended for use inside Pluto notebooks to ease composition of bigger scripts via smaller parts.

When an HTLScript is interpolated inside the @htl macro, the following code is generated:

<script id=$id>
$body

invalidation.then(() => {
	$invalidation
})

If the id = missing (default), a random string id is associated to the script. If id = nothing, a script without id is created.

If the invalidation field is missing, the whole invalidation block is skipped.

Multiple HTLScript elements can be combined together using the combine_script function also exported by this package, allowing to generate bigger scripts by composing multiple building blocks.

When shown inside the output of Pluto cells, the HTLScript object prints its containing formatted code as a Markdown.Code element.

Javascript Events Listeners

HTLScript provides some simplified way of adding event listeners in javascript that are automatically removed upon cell invalidation. Scripts created using HTLScript expose an internal javascript function

addScriptEventListener(element, listeners)

which accepts any givent JS element to which listeners have to be attached, and an object of with the following key-values:

{ 
  eventName1: listenerFunction1, 
  eventName2: listenerFunction2,
  ... 
}

When generating the script to execute, HTLScript automatically adds all the provided listeners to the provided element, and also takes care of removing all the listeners upon cell invalidation.

For example, the following julia code:

let
	script = HTLScript(@htl("""
<script>
	addScriptEventListeners(window, { 
		click: function (event) {
			console.log('click: ',event)
		}, 
		keydown: function (event) {
			console.log('keydown: ',event)
		},
	})
</script>
	"""))
	@htl"$script"
end

is functionally equivalent of writing the following javascript code within the script tag of the cell output

function onClick(event) {
	console.log('click: ',event)
}
function onKeyDown(event) {
	console.log('keydown: ',event)
}
window.addEventListener('click', onClick)
window.addEventListener('keydown', onKeyDown)

invalidation.then(() => {
	window.removeEventListener('click', onClick)
	window.removeEventListener('keydown', onKeyDown)
})

See also: HTLScriptPart, combine_scripts

Examples:

The following code:

let
a = HTLScript("console.log('asd')")
b = HTLScript(@htl("<script>console.log('boh')</script>"), "console.log('lol')")
script = combine_scripts([a,b];id="test")
out = @htl("$script")
end

is equivalent to writing directly

@htl """
<script id='test'>
	console.log('asd')
	console.log('boh')
	
	invalidation.then(() => {
		console.log('lol')
	})
"""
</script>

PlutoDevMacros.Script.HTLScriptPart — Type

struct HTLScriptPart

This struct is a simple wrapper around an IOBuffer and is intended to provide pretty printing of script contents and custom interpolation inside the <script> tags of the @htl macro.

It is intended for use within Pluto notebooks to simply decouple parts of a javascript script into separate variables and still be able to interpolate them within <script> tags to compose a bigger script.

Compared to simply using strings wrapped in HypertextLiteral.JavaScript, this gives the opportunity to exploit the experimental htmlmixed synthax highlighting of code inside cells by also accepting HypertextLiteral.Result objects during construction.

The struct can be initialized as follows:

wrapper = HTLScriptPart(@htl("""
<script>
code...
</script>
"""))

When interpolating wrapper above inside another @htl macro as @htl "<script>$wrapper</script>" it would be as equivalent to directly writing @htl "<script>code...</script> inside the script. This is clearly only beneficial if multiple HTLScriptPart variables are interpolated inside a single <script> block.

On top of the interpolation, an object of type HTLScriptPart will show its contents as a formatted javascript code markdown element when shown in Pluto.

The constructor also accepts either a String or IOBuffer object for its initialization instead of a HypertextLiteral.Result one.

See also: HTLScript, HTLBypass

Examples:

let
	asd = HTLScriptPart(@htl """
	<script>
		let out = html`<div></div>`
		console.log('first script')
	</script>
	""")
	lol = HTLScriptPart(@htl """
	<script>
		let a = Math.random()
		out.innerText = a
		console.log('second script')
		return out
	</script>
	""")
	@htl """
	<script>
		$([
			asd,
			lol
		])
	</script>
	"""
end

PlutoDevMacros.Script.HTLBypass — Type

struct HTLBypass

This struct is a simple wrapper around HypertextLiteral.Result intended to provide interpolation inside <script> tags as if writing the code that generated the result directly.

This is intended for use inside Pluto notebooks to ease variable interpolation inside html element generated within <script> tags using the html`` command that is imported from Observable.

This way, one can generate the intended HTML inside other cells to more easily see the results and with support of nested @htl interpolation.

The struct only accepts the output of the @htl macro as an input.

On top of the interpolation, an object of type HTLBypass will simply show the wrapped HypertextLiteral.Result when shown with MIME"text/html".

See also: HTLScriptPart, HTLScript

Examples:

let
	bpclass = "magic";
	bplol = @htl """
	<div class=$bpclass>
		MAGIC
	</div>
	"""
	bpasd = HTLBypass(@htl "
	<div>This is $bplol</div>
	")
	@htl "<script>let out = html`$bpasd`;return out</script>"
end

PlutoDevMacros.Script.combine_scripts — Function

combine_scripts(v; id = missing)

Creates a single HTLScript element by combining multiple inputs contained in the iterable object v.

The generated combined output has its fields body and invalidation that are the concatenation of all the body and invalidation fields (respectively) of the elements of v. Elements of v that are not of type HTLScript are converted (using the HTLScript constructor) prior to concatenation.

When a custom id is provided as kwarg, the final script id will be id, otherwise it is computed depending on the id fields of the elements of v.

See also: HTLScript

Examples:

The following code:

let
	a = HTLScript("console.log('asd')")
	b = HTLScript(@htl("<script>console.log('boh')</script>"), "console.log('lol')")
	script = combine_scripts([a,b];id='test')
	out = @htl("$script")
end

is equivalent to writing directly

@htl """
<script id='test'>
	console.log('asd')
	console.log('boh')
	
	invalidation.then(() => {
		console.log('lol')
	})
</script>
"""